+++ jest.config.js
@@ -12,6 +12,7 @@ module.exports = {
   ],
   moduleNameMapper: {
     "\\.svg$": "<rootDir>/__mocks__/svgMock.js",
+    "\\.(m4a|mp3|wav)$": "<rootDir>/__mocks__/audioAssetMock.js",
     "^expo-keep-awake$": "<rootDir>/__mocks__/expoKeepAwake.js",
     "^expo-clipboard$": "<rootDir>/__mocks__/expoClipboard.js",
     "^expo-file-system$": "<rootDir>/__mocks__/expoFileSystem.js",